+++ client/app/components/items-list/components/ItemsTable.jsx
@@ -6,7 +6,7 @@ import Table from "antd/lib/table";
 import FavoritesControl from "@/components/FavoritesControl";
 import TimeAgo from "@/components/TimeAgo";
 import { durationHumanize } from "@/filters";
-import { formatDate, formatDateTime } from "@/filters/datetime";
+import { formatDate, formatDateTime } from "@/filters";
 
 // `this` refers to previous function in the chain (`Columns.***`).
 // Adds `sorter: true` field to column definition
