--- client/app/directives/compare-to.js
@@ -1,28 +1,0 @@
-function compareTo() {
-  return {
-    require: "ngModel",
-    scope: {
-      otherModelValue: "=compareTo",
-    },
-    link(scope, element, attributes, ngModel) {
-      const validate = value => {
-        ngModel.$setValidity("compareTo", value === scope.otherModelValue);
-      };
-
-      scope.$watch("otherModelValue", () => {
-        validate(ngModel.$modelValue);
-      });
-
-      ngModel.$parsers.push(value => {
-        validate(value);
-        return value;
-      });
-    },
-  };
-}
-
-export default function init(ngModule) {
-  ngModule.directive("compareTo", compareTo);
-}
-
-init.init = true;
