--- client/app/directives/resize-event.js
@@ -1,19 +1,0 @@
-import resizeObserver from "@/services/resizeObserver";
-
-function resizeEvent() {
-  return {
-    restrict: "A",
-    link($scope, $element, attrs) {
-      const unwatch = resizeObserver($element[0], () => {
-        $scope.$evalAsync(attrs.resizeEvent);
-      });
-      $scope.$on("$destroy", unwatch);
-    },
-  };
-}
-
-export default function init(ngModule) {
-  ngModule.directive("resizeEvent", resizeEvent);
-}
-
-init.init = true;
